--- README.md
@@ -1,8 +1,0 @@
-# Shimmy
-## Sit less, move more.
----
-**Team**: Irene, Sam, Peter, Christian
-
-An upcoming React Native mobile to help individuals break up long periods of sitting and increase overall physical activity in an enjoyable way!
-
-Stay tuned...!
