+++ package.json
@@ -5,40 +5,18 @@
     "android": "expo start --android",
     "ios": "expo start --ios",
     "web": "expo start --web",
-    "eject": "expo eject",
-    "test": "jest --watchAll"
-  },
-  "jest": {
-    "preset": "jest-expo"
+    "eject": "expo eject"
   },
   "dependencies": {
-    "@expo/vector-icons": "^12.0.0",
-    "@react-navigation/bottom-tabs": "^6.0.5",
-    "@react-navigation/native": "^6.0.2",
-    "@react-navigation/native-stack": "^6.1.0",
     "expo": "~42.0.1",
-    "expo-asset": "~8.3.2",
-    "expo-constants": "~11.0.1",
-    "expo-font": "~9.2.1",
-    "expo-linking": "~2.3.1",
-    "expo-splash-screen": "~0.11.2",
     "expo-status-bar": "~1.0.4",
-    "expo-web-browser": "~9.2.0",
     "react": "16.13.1",
     "react-dom": "16.13.1",
     "react-native": "https://github.com/expo/react-native/archive/sdk-42.0.0.tar.gz",
-    "react-native-reanimated": "~2.2.0",
-    "react-native-gesture-handler": "~1.10.2",
-    "react-native-safe-area-context": "3.2.0",
-    "react-native-screens": "~3.4.0",
     "react-native-web": "~0.13.12"
   },
   "devDependencies": {
-    "@babel/core": "^7.9.0",
-    "@types/react": "~16.9.35",
-    "@types/react-native": "~0.63.2",
-    "jest-expo": "~41.0.0-beta.0",
-    "typescript": "~4.0.0"
+    "@babel/core": "^7.9.0"
   },
   "private": true
 }
